virtual_service/requirements.txt
fastapi[standard]==0.115.4
typing==3.7.4.3
pydantic==2.9.2
uvicorn==0.32.0
python-dotenv==1.0.1
httpx==0.27.2
python-logstash==0.4.8